+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+## UNRELEASED
+
+* Issue 323: Add command line flag to manage output verbosity. Note: no equivalent
+  option was added to `ini` file.
+
 ## v1.17.0 (2026-08-18)
 
 * Minor version release for Issue 322 changes.